Re-enable probing of ucm
UCM files can have a JackControls property in a device definition. This
property defines an ALSA device name which holds the plugged or
unplugged state of a jack.

The only place where a JackControl property is mapped to a jack is in
ucm_probe_profile_set. If it is skipped, jacks are never mapped and
Pulse never recognizes when it is plugged in or unplugged.